Fix problem of int RLPDecoding in solidity

+++ solidity/bmc/contracts/libraries/RLPDecode.sol
@@ -201,7 +201,7 @@ library RLPDecode {
 
     function toInt(RLPItem memory item) internal pure returns (int256) {
         if ((toBytes(item)[0] & 0x80) == 0x80) {
-            return int256(toUint(item) - 2**(toBytes(item).length * 8));
+            return int256(toUint(item)) - int256(2**(toBytes(item).length * 8));
         }
 
         return int256(toUint(item));
